8.18 修改副本的Boss技能施放机制，新增多职业的技能，新增Boss的多技能施放

+++ src/main/java/com/ljh/gamedemo/entity/Creep.java
@@ -13,7 +13,7 @@ public class Creep {
     /**
      * 野怪id
      */
-    private Integer creepId;
+    private Long creepId;
 
     /**
      * 野怪类型
